requirements.txt
beautifulsoup4==4.8.0
numpy==1.18.4
pandas==0.25.1
requests==2.22.0
psycopg2-binary==2.8.4
